refactor: inline gateway/proxy/egress-judge index re-export
The index.ts was a 6-line pass-through re-exporting EgressJudge from
./judge and JudgeClient/JudgeDecision/JudgeVerdict types from ./types.
Point the four importers (http-proxy.ts plus two test files) directly
at judge.js / types.js and delete the wrapper.
