--- packages/server/src/gateway/proxy/egress-judge/index.ts
@@ -1,6 +1,0 @@
-export { EgressJudge } from "./judge.js";
-export type {
-  JudgeClient,
-  JudgeDecision,
-  JudgeVerdict,
-} from "./types.js";
